refactor folder structure for components
