+++ src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
-import { ScrollToTopComponent } from "./components/scroll-to-top/scroll-to-top.component";
+import { ScrollToTopComponent } from "./components/shared/scroll-to-top/scroll-to-top.component";
 
 @Component({
   selector: 'app-root',
