+++ src/app/components/admin/gym-config/gym-config.component.html
