+++ src/app/components/management/user-details/user-details.component.css
