+++ src/app/components/management/user-form/user-form.component.ts
@@ -1,18 +1,18 @@
 import { Component, inject, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
 import { ActivatedRoute, Router, RouterModule } from '@angular/router';
-import { User } from '../../models/users/users.interface';
-import { Gender } from '../../models/users/gender.interface';
-import { Eps } from '../../models/users/eps.interface';
-import { BloodType } from '../../models/users/bloodtype.interface';
-import { BloodRh } from '../../models/users/bloodrh.interface';
+import { User } from '../../../models/users/users.interface';
+import { Gender } from '../../../models/users/gender.interface';
+import { Eps } from '../../../models/users/eps.interface';
+import { BloodType } from '../../../models/users/bloodtype.interface';
+import { BloodRh } from '../../../models/users/bloodrh.interface';
 import { NgFor, NgIf } from '@angular/common';
-import { ageValidator } from '../../validators/ageValidator';
-import { UserService } from '../../services/users/users.service';
-import { GenderService } from '../../services/users/gender.service';
-import { EpsService } from '../../services/users/eps.service';
-import { BloodTypeService } from '../../services/users/bloodtype.service';
-import { BloodRhService } from '../../services/users/bloodrh.service';
+import { ageValidator } from '../../../validators/ageValidator';
+import { UserService } from '../../../services/users/users.service';
+import { GenderService } from '../../../services/users/gender.service';
+import { EpsService } from '../../../services/users/eps.service';
+import { BloodTypeService } from '../../../services/users/bloodtype.service';
+import { BloodRhService } from '../../../services/users/bloodrh.service';
 
 @Component({
   selector: 'app-user-form',
