+++ src/app/pages/home/home.component.ts
@@ -2,7 +2,7 @@ import { Component, inject, Inject, OnInit } from '@angular/core';
 import { LoginService } from '../../services/auth/login.service';
 import { CommonModule } from '@angular/common';
 import { RouterLink } from '@angular/router';
-import HomeNavComponent from "../../components/home-nav/home-nav.component";
+import HomeNavComponent from "../../components/shared/home-nav/home-nav.component";
 
 @Component({
   selector: 'app-home',
