+++ src/app/pages/plans/plans.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
-import HomeNavComponent from "../../components/home-nav/home-nav.component";
-import MembershipTypeListComponent from "../../components/membership-type-list/membership-type-list.component";
+import HomeNavComponent from "../../components/shared/home-nav/home-nav.component";
+import MembershipTypeListComponent from "../../components/shared/membership-type-list/membership-type-list.component";
 
 @Component({
   selector: 'app-plans',
